fix: address CodeRabbit review feedback and fix failing test
- Update Pydantic schema duration_seconds from int to float to match
  the DB column type change (animation.py)
- Consolidate storage age bucket queries into a single SQL statement
  using CASE expressions to prevent race conditions and negative
  bucket values from concurrent inserts (stats.py)
- Remove redundant local `import os` (already at module level) and log
  debug message on PermissionError instead of silently swallowing
  (image_operations.py)
- Add fps=0 guard on frame_duration fallback to prevent ZeroDivisionError
  in both _create_initial_video and _create_ffmpeg_video (video_handler.py)
- Update test_pingpong_two_frames to expect correct [1,2,1] bounce
  behavior instead of codifying the old bug (test_animation_helpers.py)

All 1632 backend tests pass.

+++ backend/app/models/animation.py
@@ -106,7 +106,7 @@ class AnimationResponse(BaseModel):
     scale: str = "100%"
     output_path: str | None = None
     file_size: int = 0
-    duration_seconds: int = 0
+    duration_seconds: float = 0.0
     created_at: datetime | None = None
     completed_at: datetime | None = None
     error: str = ""
